requirements.txt
bcrypt==4.1.3
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
dash==4.1.0
Flask==3.0.3
Flask-Cors==4.0.1
idna==3.11
importlib_metadata==9.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
narwhals==2.18.1
nest-asyncio==1.6.0
numpy==2.4.4
packaging==26.0
pandas==3.0.1
plotly==6.6.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.33.1
retrying==1.4.2
setuptools==82.0.1
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.7
zipp==3.23.0
